binder: remove dead logging code
This code is only enabled if a static constant is set to false.
However, there are other (and better) ways to track transactions
these days.

For instance, you can use perfetto. There are also alternatives
like the Java-only observers.

Also, the fact that this is only on the main thread is really
arbitrary and useless, and why is there custom code here anyway
implementing the eventlog protocol?

Bug: 290731071
Test: boot (change should be no-op)
Google: 2654297
Change-Id: Ic7d7e3d045f9c7f53fc576bf2237ebe32a5a7f92